Add business rule that validates user on deleting events or expenses. Only creator or parent creator are allowed to delete.

+++ backend/Expenses.Application/Features/Events/Commands/DeleteEvent/DeleteEventCommand.cs
@@ -1,4 +1,5 @@
-﻿using Expenses.Application.Common.Interfaces;
+﻿using Expenses.Application.Common.Exceptions;
+using Expenses.Application.Common.Interfaces;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using System.Threading;
@@ -14,10 +15,12 @@ namespace Expenses.Application.Features.Events.Commands.DeleteEvent
     public class DeleteEventCommandHandler : IRequestHandler<DeleteEventCommand>
     {
         private readonly IAppDbContext _context;
+        private readonly ICurrentUserService _currentUserService;
 
-        public DeleteEventCommandHandler(IAppDbContext context)
+        public DeleteEventCommandHandler(IAppDbContext context, ICurrentUserService currentUserService)
         {
             _context = context;
+            _currentUserService = currentUserService;
         }
 
         public async Task<Unit> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
@@ -25,6 +28,10 @@ namespace Expenses.Application.Features.Events.Commands.DeleteEvent
             var @event = await _context.Events
                 .Include(e => e.Expenses)
                 .SingleOrDefaultAsync(e => e.Id == request.Id);
+
+            // Businessrule: Only allowed for expense creator or event creator
+            if (@event.Creator.Id != _currentUserService.UserId)
+                throw new ForbiddenAccessException();
 
             _context.Events.Remove(@event);
             await _context.SaveChangesAsync(cancellationToken);
